+++ packages/client/src/fetchMetadata.test.ts
@@ -307,6 +307,7 @@ describe("FetchMetadata", () => {
         "rid": "ri.ontology.main.action-type.9f84017d-cf17-4fa8-84c3-8e01e5d594f2",
         "status": "ACTIVE",
         "type": "action",
+        "unsanitizedApiName": "moveOffice",
       }
     `);
   });