+++ packages/generator/src/v2.0/generateClientSdkVersionTwoPointZero.test.ts
@@ -643,17 +643,20 @@ describe("generator", () => {
             rid: 'ri.ontology.main.action-type.8f94017d-cf17-4fa8-84c3-8e01e5d594f2';
             status: 'ACTIVE';
             type: 'action';
+            unsanitizedApiName: 'deleteTodos';
 
             signatures: deleteTodos.Signatures;
           };
           apiName: 'deleteTodos';
           type: 'action';
+          unsanitizedApiName: 'deleteTodos';
           osdkMetadata: typeof $osdkMetadata;
         }
 
         export const deleteTodos: deleteTodos = {
           apiName: 'deleteTodos',
           type: 'action',
+          unsanitizedApiName: 'deleteTodos',
           osdkMetadata: $osdkMetadata,
         };
         ",
@@ -729,17 +732,20 @@ describe("generator", () => {
             rid: 'ri.ontology.main.action-type.9f84017d-cf17-4fa8-84c3-8e01e5d594f2';
             status: 'ACTIVE';
             type: 'action';
+            unsanitizedApiName: 'markTodoCompleted';
 
             signatures: markTodoCompleted.Signatures;
           };
           apiName: 'markTodoCompleted';
           type: 'action';
+          unsanitizedApiName: 'markTodoCompleted';
           osdkMetadata: typeof $osdkMetadata;
         }
 
         export const markTodoCompleted: markTodoCompleted = {
           apiName: 'markTodoCompleted',
           type: 'action',
+          unsanitizedApiName: 'markTodoCompleted',
           osdkMetadata: $osdkMetadata,
         };
         ",
@@ -1327,17 +1333,20 @@ describe("generator", () => {
             rid: 'ri.ontology.main.action-type.8f94017d-cf17-4fa8-84c3-8e01e5d594f2';
             status: 'ACTIVE';
             type: 'action';
+            unsanitizedApiName: 'foo.bar.deleteTodos';
 
             signatures: deleteTodos.Signatures;
           };
           apiName: 'foo.bar.deleteTodos';
           type: 'action';
+          unsanitizedApiName: 'foo.bar.deleteTodos';
           osdkMetadata: typeof $osdkMetadata;
         }
 
         export const deleteTodos: deleteTodos = {
           apiName: 'foo.bar.deleteTodos',
           type: 'action',
+          unsanitizedApiName: 'foo.bar.deleteTodos',
           osdkMetadata: $osdkMetadata,
         };
         ",
@@ -1413,17 +1422,20 @@ describe("generator", () => {
             rid: 'ri.ontology.main.action-type.9f84017d-cf17-4fa8-84c3-8e01e5d594f2';
             status: 'ACTIVE';
             type: 'action';
+            unsanitizedApiName: 'foo.bar.markTodoCompleted';
 
             signatures: markTodoCompleted.Signatures;
           };
           apiName: 'foo.bar.markTodoCompleted';
           type: 'action';
+          unsanitizedApiName: 'foo.bar.markTodoCompleted';
           osdkMetadata: typeof $osdkMetadata;
         }
 
         export const markTodoCompleted: markTodoCompleted = {
           apiName: 'foo.bar.markTodoCompleted',
           type: 'action',
+          unsanitizedApiName: 'foo.bar.markTodoCompleted',
           osdkMetadata: $osdkMetadata,
         };
         ",
@@ -2239,17 +2251,20 @@ describe("generator", () => {
               rid: 'ri.a.b.c.d';
               status: 'ACTIVE';
               type: 'action';
+              unsanitizedApiName: 'setTaskBody';
 
               signatures: setTaskBody.Signatures;
             };
             apiName: 'setTaskBody';
             type: 'action';
+            unsanitizedApiName: 'setTaskBody';
             osdkMetadata: typeof $osdkMetadata;
           }
 
           export const setTaskBody: setTaskBody = {
             apiName: 'setTaskBody',
             type: 'action',
+            unsanitizedApiName: 'setTaskBody',
             osdkMetadata: $osdkMetadata,
           };
           "