+++ packages/react-components/src/filter-list/inputs/NumberRangeFilterInput.tsx
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-import type { ObjectTypeDefinition } from "@osdk/api";
+import type { ObjectSet, ObjectTypeDefinition } from "@osdk/api";
 import { useOsdkAggregation } from "@osdk/react/experimental";
 import React, { memo, useCallback, useMemo } from "react";
 import { NullValueWrapper } from "../base/inputs/NullValueWrapper.js";
@@ -28,6 +28,7 @@ import {
 
 interface NumberRangeFilterInputProps<Q extends ObjectTypeDefinition> {
   objectType: Q;
+  objectSet: ObjectSet<Q>;
   propertyKey: string;
   filterState: FilterState | undefined;
   onFilterStateChanged: (state: FilterState) => void;
@@ -35,6 +36,7 @@ interface NumberRangeFilterInputProps<Q extends ObjectTypeDefinition> {
 
 function NumberRangeFilterInputInner<Q extends ObjectTypeDefinition>({
   objectType,
+  objectSet,
   propertyKey,
   filterState,
   onFilterStateChanged,
@@ -78,8 +80,8 @@ function NumberRangeFilterInputInner<Q extends ObjectTypeDefinition>({
   );
 
   const histogramArgs = useMemo(
-    () => ({ aggregate: aggregateOptions }),
-    [aggregateOptions],
+    () => ({ aggregate: aggregateOptions, objectSet }),
+    [aggregateOptions, objectSet],
   );
 
   const { data: aggregateData, isLoading: histLoading } = useOsdkAggregation(
@@ -120,8 +122,12 @@ function NumberRangeFilterInputInner<Q extends ObjectTypeDefinition>({
   );
 
   const nullCountArgs = useMemo(
-    () => ({ where: nullWhereClause, aggregate: nullCountAggregateOptions }),
-    [nullWhereClause, nullCountAggregateOptions],
+    () => ({
+      where: nullWhereClause,
+      aggregate: nullCountAggregateOptions,
+      objectSet,
+    }),
+    [nullWhereClause, nullCountAggregateOptions, objectSet],
   );
 
   const {
@@ -142,17 +148,19 @@ function NumberRangeFilterInputInner<Q extends ObjectTypeDefinition>({
     return 0;
   }, [nullCountData]);
 
+  const isLoading = histLoading || nullLoading;
+
   return (
     <NullValueWrapper
       nullCount={nullCount}
-      isLoading={nullLoading}
+      isLoading={isLoading}
       error={nullError}
       includeNull={includeNull}
       onIncludeNullChange={handleNullChange}
     >
       <NumberRangeInput
         valueCountPairs={valueCountPairs}
-        isLoading={histLoading}
+        isLoading={isLoading}
         minValue={numberRangeState?.minValue}
         maxValue={numberRangeState?.maxValue}
         onChange={handleRangeChange}