+++ benchmarks/tests/primary/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @osdk/benchmarks.primary
 
+## 0.1.0-beta.124
+
+### Patch Changes
+
+- @osdk/client@2.8.0-beta.25
+
 ## 0.1.0-beta.123
 
 ### Patch Changes